agents/the-underwriter/pyproject.toml
[tool.poetry]
name = "the-underwriter"
version = "0.1.0"
description = "Agent insurance, guarantees, and reputation — the trust layer for the agent economy"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
payments-py = {version = ">=1.1.0", extras = ["mcp"]}
fastapi = "^0.120.0"
uvicorn = ">=0.34.2,<1.0.0"
httpx = "^0.28.0"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
